lmb-1214: can route to base page for detail of form
