+++ app/router.js
@@ -67,6 +67,8 @@ Router.map(function () {
 
   this.route('eigen-gegevens', function () {
     this.route('new');
+    this.route('detail');
+    this.route('overzicht');
   });
 
   this.route('under-construction');
