+++ app/routes/eigen-gegevens.js
@@ -1,20 +1,16 @@
 import Route from '@ember/routing/route';
 
-import moment from 'moment';
+import { service } from '@ember/service';
 
 export default class EigenGegevensRoute extends Route {
-  async model() {
-    return {
-      forms: [
-        {
-          label: 'test formulier',
-          description: `
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
-          `,
-          createdAt: moment(new Date()).format('DD-MM-YYYY HH:mm'),
-          modified: moment(new Date()).format('DD-MM-YYYY HH:mm'),
-        },
-      ],
-    };
+  @service session;
+  @service router;
+
+  beforeModel(transition) {
+    this.session.requireAuthentication(transition, 'login');
+  }
+
+  model() {
+    this.router.transitionTo('eigen-gegevens.overzicht');
   }
 }
