+++ tests/unit/routes/eigen-gegevens/detail-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'frontend-lmb/tests/helpers';
+
+module('Unit | Route | eigen-gegevens/detail', function (hooks) {
+  setupTest(hooks);
+
+  test('it exists', function (assert) {
+    let route = this.owner.lookup('route:eigen-gegevens/detail');
+    assert.ok(route);
+  });
+});
