+++ tests/unit/routes/eigen-gegevens/overzicht-test.js
@@ -1,0 +1,11 @@
+import { module, test } from 'qunit';
+import { setupTest } from 'frontend-lmb/tests/helpers';
+
+module('Unit | Route | eigen-gegevens/overzicht', function (hooks) {
+  setupTest(hooks);
+
+  test('it exists', function (assert) {
+    let route = this.owner.lookup('route:eigen-gegevens/overzicht');
+    assert.ok(route);
+  });
+});
